The software license has been updated to MPL 2.0
The reason why the software is licensed is to prevent malicious people from commercially feeding on the software and to ensure that every publicly made change is officially shared.
Laniatus Games Studio Inc. As a company, we have never tended to be profit-oriented, and we will not allow others to follow this trend over our craft.
We always choose to stick to our vision.

@Laniatus Games Studio Inc.